Uusin versio.

+++ app/src/main/java/com/example/bankapplication/CreditAccount.java
@@ -2,17 +2,18 @@ package com.example.bankapplication;
 
 public class CreditAccount extends Account {
 
-    private int credit;
-    public CreditAccount(String eM, String aN, int b, int c) {
+    private float credit;
+
+    public CreditAccount(String eM, String aN, float b, float c) {
         super(eM, aN, b);
         credit = c;
     }
 
-    public int getCreditLimit() {
+    public float getCreditLimit() {
         return credit;
     }
 
-    public void setBalance(int amount) {
+    public void setBalance(float amount) {
         if (-credit <= balance + amount)
             this.balance += amount;
