+++ app/src/main/java/com/example/bankapplication/RegularAccount.java
@@ -2,7 +2,7 @@ package com.example.bankapplication;
 
 public class RegularAccount extends Account {
 
-    public RegularAccount(String email, String account_number, int balance) {
+    public RegularAccount(String email, String account_number, float balance) {
         super(email, account_number, balance);
     }
